PF/backend/requirements.txt
Django~=4.1.3
Pillow~=9.3.0
django-recurrence~=1.11.1
djangorestframework~=3.14.0
geopy~=2.3.0
phonenumbers~=8.13.0
djangorestframework-simplejwt~=5.2.2
django-cors-headers~=3.13.0